+++ SQLWATCHDB/dbo/Tables/sqlwatch_config_sql_instance.sql
@@ -2,7 +2,7 @@
 (
 	[sql_instance] varchar(32) not null constraint df_sqlwatch_config_sql_instance_sql_instance default (@@SERVERNAME),
 	[hostname] nvarchar(32) null,
-	[sql_port] smallint null,
+	[sql_port] int null,
 	[sqlwatch_database_name] sysname not null constraint df_sqlwatch_config_sql_instance_database_name default (DB_NAME()),
 	[environment] sysname not null constraint df_sqlwatch_config_sql_instance_env default ('DEFAULT'),
 	[repo_collector_is_active] bit not null constraint df_sqlwatch_config_sql_instance default (1),
